feat(user-join-screen): move queries to a exclusive file
